requirements.txt
Flask==3.0.3
gunicorn==20.1.0
SQLAlchemy==2.0.18
requests==2.31.0
python-dotenv==1.0.1